renamed classes; replaced AppFactory with the new builder.
